move sparse array into its own file
